Extract common counter UI
